Revert "ACZ fixes (#1573)"
This reverts commit 173ca011fdc4e0725ae00303746466220579d7e4.
