+++ OpenDreamRuntime/DreamManager.cs
@@ -116,7 +116,7 @@ namespace OpenDreamRuntime {
             }
 
             _compiledJson = json;
-            var rootPath = Path.GetFullPath(Path.GetDirectoryName(jsonPath)!);
+            var rootPath = Path.GetDirectoryName(jsonPath)!;
             var resources = _compiledJson.Resources ?? Array.Empty<string>();
             _dreamResourceManager.Initialize(rootPath, resources);
             if(!string.IsNullOrEmpty(_compiledJson.Interface) && !_dreamResourceManager.DoesFileExist(_compiledJson.Interface))
@@ -146,9 +146,9 @@ namespace OpenDreamRuntime {
 
             _dreamMapManager.LoadMaps(_compiledJson.Maps);
 
-            var aczProvider = new DreamAczProvider(_dependencyCollection, rootPath, resources);
-            _statusHost.SetMagicAczProvider(aczProvider);
-            _statusHost.SetFullHybridAczProvider(aczProvider);
+            _statusHost.SetMagicAczProvider(new DreamMagicAczProvider(
+                _dependencyCollection, rootPath, resources
+            ));
 
             return true;
         }
